requirements.txt

requests

# testing
mock==3.0.5
httpretty==1.1.4
coveralls==1.11.1
coverage==5.2.1
google-api-python-client==2.24.0
python-dotenv
numpy
pandas
setuptools